Re-structure package, move modules to top level
