+++ src/isr_spectrum/__init__.py
@@ -1,5 +1,7 @@
-from isr_spectrum.utils.config import *  # noqa: F401, F403
-from isr_spectrum.utils.integrand_functions import *  # noqa: F401, F403
-from isr_spectrum.utils.spectrum_calculation import SpectrumCalculation  # noqa: F401
+"""Initialize the isr_spectrum package."""
+
+from isr_spectrum.config import *  # noqa: F401, F403
+from isr_spectrum.integrand_functions import *  # noqa: F401, F403
+from isr_spectrum.spectrum_calculation import SpectrumCalculation  # noqa: F401
 
 __version__ = "3.0.0"
